requirements.txt
Flask
pdfplumber
pytesseract
pdf2image
python-docx
pandas
gunicorn
